+++ templates/integrated/Source/Entities/Member.cs
@@ -5,12 +5,12 @@ sealed class Member : Entity
     public ulong MemberNumber { get; set; }
     public DateOnly SignupDate { get; set; }
 
-    public string FirstName { get; set; }
-    public string LastName { get; set; }
+    public string FirstName { get; init; }
+    public string LastName { get; init; }
     public string Designation { get; set; }
-    public string Email { get; set; }
+    public string Email { get; init; }
     public DateOnly BirthDay { get; set; }
     public string Gender { get; set; }
-    public string MobileNumber { get; set; }
+    public string MobileNumber { get; init; }
     public Address Address { get; set; }
 }