+++ templates/integrated/Source/Entities/NotificationTemplate.cs
@@ -5,11 +5,11 @@ namespace Dom;
 sealed class NotificationTemplate : IEntity
 {
     [BsonId]
-    public string ID { get; set; } //set the template name as id
+    public string ID { get; init; } //set the template name as id
 
-    public string SmsBody { get; set; }
-    public string EmailSubject { get; set; }
-    public string EmailBody { get; set; }
+    public string SmsBody { get; init; }
+    public string EmailSubject { get; init; }
+    public string EmailBody { get; init; }
 
     public object GenerateNewID()
         => ID; //because we're setting the ID manually