+++ templates/integrated/Source/Features/Members/SignUp/Request.cs
@@ -7,22 +7,22 @@ sealed class Request
 {
     //request dto shape is dictated by front-end team
 
-    public User UserDetails { get; set; }
-    public string Email { get; set; }
-    public string BirthDay { get; set; }
-    public string Gender { get; set; }
-    public ContactDetails Contact { get; set; }
-    public AddressDetails Address { get; set; }
+    public User UserDetails { get; init; }
+    public string Email { get; init; }
+    public string BirthDay { get; init; }
+    public string Gender { get; init; }
+    public ContactDetails Contact { get; init; }
+    public AddressDetails Address { get; init; }
 
     public sealed class User
     {
-        public string FirstName { get; set; }
-        public string LastName { get; set; }
+        public string FirstName { get; init; }
+        public string LastName { get; init; }
     }
 
     public sealed class ContactDetails
     {
-        public string MobileNumber { get; set; }
+        public string MobileNumber { get; init; }
         public bool Whatsapp { get; set; }
         public bool Viber { get; set; }
         public bool Telegram { get; set; }
@@ -30,13 +30,13 @@ sealed class Request
 
     public sealed class AddressDetails
     {
-        public string Street { get; set; }
-        public string City { get; set; }
-        public string State { get; set; }
-        public string ZipCode { get; set; }
+        public string Street { get; init; }
+        public string City { get; init; }
+        public string State { get; init; }
+        public string ZipCode { get; init; }
     }
 
-    sealed class Validator : Validator<Request>
+    internal sealed class Validator : Validator<Request>
     {
         public Validator()
         {