+++ templates/project/Source/Features/SayHello/Request.cs
@@ -4,8 +4,8 @@ namespace SayHello;
 
 sealed class Request
 {
-    public string FirstName { get; set; }
-    public string LastName { get; set; }
+    public string FirstName { get; init; }
+    public string LastName { get; init; }
 
     internal sealed class Validator : Validator<Request>
     {