+++ templates/project/Source/Program.cs
@@ -1,10 +1,17 @@
+using FastEndpoints.Security;
+using MyProject;
+
 var bld = WebApplication.CreateBuilder(args);
 bld.Services
-   .AddFastEndpoints()
+   .AddAuthenticationJwtBearer(s => s.SigningKey = bld.Configuration["Auth:JwtKey"])
+   .AddAuthorization()
+   .AddFastEndpoints(o => o.SourceGeneratorDiscoveredTypes.AddRange(DiscoveredTypes.All))
    .SwaggerDocument();
 
 var app = bld.Build();
-app.UseFastEndpoints()
+app.UseAuthentication()
+   .UseAuthorization()
+   .UseFastEndpoints(c => c.Errors.UseProblemDetails())
    .UseSwaggerGen();
 app.Run();